requirements.txt
comet-ml==3.33.9
torch==2.1.2
librosa==0.10.1
matplotlib==3.7.2
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
torchsummary==1.5.1
tqdm==4.66.1
heartpy==1.2.7
fastdtw==0.3.4
seaborn==0.12.2
neurokit2
flask
joblib
pyampd
mat73 